refactor(collision): :recycle: refactor core package to collisions as it only have collisions
It is only a refactor, no logic change. Only the package name is changed.
