+++ duelinvaders/src/main/java/org/enstabretagne/Collision/BulletBulletCollision.java
@@ -1,4 +1,4 @@
-package org.enstabretagne.Core;
+package org.enstabretagne.Collision;
 
 import static com.almasb.fxgl.dsl.FXGL.play;
 import static com.almasb.fxgl.dsl.FXGL.spawn;
