+++ duelinvaders/src/test/java/org/enstabretagne/Core/AlienBulletCollisionTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import com.almasb.fxgl.entity.Entity;
 import com.almasb.fxgl.entity.GameWorld;
 
+import org.enstabretagne.Collision.AlienBulletCollision;
 // import static org.junit.jupiter.api.Assertions.*;
 import org.enstabretagne.Component.*;
 import org.enstabretagne.Utils.EntityType;
